backend/requirements.txt
tiktoken==0.5.1
watchdog==3.0.0
litellm==1.17.5
fastapi==0.103.2
uvicorn==0.25.0
pydantic==2.4.2
pydantic-settings==2.0.3
beautifulsoup4==4.12.2
selenium==4.14.0
webdriver-manager==4.0.1
pillow==10.0.1
tomlkit==0.12.1
youtube-transcript-api==0.6.1
python-multipart==0.0.6
appdirs==1.4.4
rtoml==0.10.0
send2trash==1.8.2
email-validator==2.1.0.post1
deepmerge==1.1.1
sentry-sdk[fastapi]==1.39.1
rich==13.7.0
types-requests==2.31.0.20240106
types-setuptools==69.0.0.20240106
platformdirs==4.1.0
pytest-repeat==0.9.3
types-send2trash==1.8.2.7
jupyter-client==8.6.0
ipykernel==6.29.2
matplotlib==3.8.2
